+++ coredis/client/basic.py
@@ -648,7 +648,6 @@ class Redis(Client[AnyStr]):
         Changes
           - .. versionremoved:: 6.0.0
             - :paramref:`protocol_version` removed (and therefore support for RESP2)
-
           - .. versionadded:: 6.0.0
             -  TODO: Add stuff
           - .. versionadded:: 4.12.0