+++ coredis/commands/_key_spec.py
@@ -489,14 +489,6 @@ class KeySpec:
         b"FT.SUGDEL": lambda args: (args[1],),
         b"FT.SUGGET": lambda args: (args[1],),
         b"FT.SUGLEN": lambda args: (args[1],),
-        b"GRAPH.QUERY": lambda args: (args[1],),
-        b"GRAPH.DELETE": lambda args: (args[1],),
-        b"GRAPH.EXPLAIN": lambda args: (args[1],),
-        b"GRAPH.PROFILE": lambda args: (args[1],),
-        b"GRAPH.SLOWLOG": lambda args: (args[1],),
-        b"GRAPH.CONSTRAINT CREATE": lambda args: (args[1],),
-        b"GRAPH.CONSTRAINT DROP": lambda args: (args[1],),
-        b"GRAPH.RO_QUERY": lambda args: (args[1],),
     }
 
     @classmethod